fix: share fetch boundary error type
