+++ apps/api/src/lib/errors/tagged-errors.ts
@@ -2,6 +2,8 @@ import { TaggedError } from "better-result";
 
 import type { ChatTransportErrorCode } from "@stll/anonymize-chat";
 
+export { FetchBoundaryError } from "@stll/errors";
+
 export type HandlerErrorStatusCode =
   | 400
   | 401
@@ -195,19 +197,6 @@ export class ExtractionWorkerError extends TaggedError(
 )<{
   message: string;
   exitCode: number | null;
-}>() {}
-
-/**
- * HTTP/network failure at an outbound or inbound boundary. Carries the URL,
- * status, and optional response body for structured logging and retry logic.
- */
-export class FetchBoundaryError extends TaggedError("FetchBoundaryError")<{
-  url: string;
-  status?: number;
-  statusText?: string;
-  body?: string;
-  message: string;
-  cause?: unknown;
 }>() {}
 
 /** Timeout waiting for a readiness probe, subprocess, or external resource. */
