+++ apps/web/src/lib/errors/index.ts
@@ -1,5 +1,7 @@
 import { TaggedError } from "better-result";
 
+export { FetchBoundaryError } from "@stll/errors";
+
 export class APIError extends TaggedError("ApiError")<{
   status: number;
   message: string;
@@ -23,20 +25,6 @@ export { ClientTelemetryError } from "@/lib/errors/telemetry";
 
 export class ClientUnknownError extends TaggedError("ClientUnknownError")<{
   message: string;
-}>() {}
-
-/**
- * HTTP/network failure at an outbound boundary (raw fetch in the browser).
- * Carries the URL, status, and optional response body so callers can
- * distinguish protocol failures from app-level API errors.
- */
-export class FetchBoundaryError extends TaggedError("FetchBoundaryError")<{
-  url: string;
-  status?: number;
-  statusText?: string;
-  body?: string;
-  message: string;
-  cause?: unknown;
 }>() {}
 
 type ToAPIErrorProps = {
