+++ apps/web/tsconfig.json
@@ -7,6 +7,7 @@
       "@/api/*": ["../api/src/*"], // this is for eden to work https://elysiajs.com/eden/installation#path-alias-monorepo
       "@stll/ares": ["../../packages/ares/src/index.ts"],
       "@stll/docx-core": ["../../packages/docx-core/src/index.ts"],
+      "@stll/errors": ["../../packages/errors/src/index.ts"],
       "@stll/infosoud": ["../../packages/infosoud/src/index.ts"],
       "@stll/money": ["../../packages/money/src/index.ts"],
       "@stll/permissions": ["../../packages/permissions/src/index.ts"],
